Update UIView extension
