--- Client/Utils/View.swift
@@ -1,41 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0
-
-import Foundation
-import UIKit
-
-extension UIView {
-    /// Convenience function to ease creating new views.
-    ///
-    /// Calling this function creates a new view with `translatesAutoresizingMaskIntoConstraints`
-    /// set to false. Passing in an optional closure to do further configuration of the view.
-    ///
-    /// - Parameter builder: A function that takes the newly created view.
-    ///
-    /// Usage:
-    /// ```
-    ///    private let button: UIButton = .build { button in
-    ///        button.setTitle("Tap me!", for state: .normal)
-    ///        button.backgroundColor = .systemPink
-    ///    }
-    /// ```
-    static func build<T: UIView>(_ builder: ((T) -> Void)? = nil) -> T {
-        let view = T()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        builder?(view)
-
-        return view
-    }
-
-    /// Convenience function to add multiple subviews
-    /// - Parameter views: A variadic parameter taking in a list of views to be added.
-    ///
-    /// Usage:
-    /// ```
-    ///    view.addSubviews(headerView, contentView, footerView)
-    /// ```
-    func addSubviews(_ views: UIView...) {
-        views.forEach(addSubview)
-    }
-}
